requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
python-dotenv
groq
